refactor: 입력에 대한 과정을 `Input` 클래스로 분리
